+++ derive_builder/CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- Forward `allow` and `cfg` attributes from the deriving struct to the builder and its impl block #222
+
 ## [0.11.0] - 2022-03-15
 - Support shorthand and long-form collection setters; `#[builder(setter(each = "..."))]` and `#[builder(setter(each(name = "...")))]` #234
 - Allow collection setters to be generic over `Into` using `#[builder(setter(each(name = "...", into)))] #234 and #214